backend/requirements.txt
fastapi>=0.109.0
sqlmodel>=0.0.16
uvicorn>=0.27.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pyjwt[crypto]>=2.8.0
httpx>=0.27.0
email-validator>=2.1.0
